hotfix(Navigation): replace Loading screen with Feed screen and update navigation parameters
