+++ src/screens/NavigationStack.tsx
@@ -3,7 +3,7 @@ import {
   StaticParamList,
 } from '@react-navigation/native';
 import { createNativeStackNavigator } from '@react-navigation/native-stack';
-import Home from './public/Home';
+import Feed from './public/Feed';
 import NotFound from './public/NotFound';
 import PhoneConfirmation from './public/PhoneConfirmation';
 import ConfirmPhoneNumber from './public/ConfirmPhoneNumber';
@@ -11,7 +11,7 @@ import RegisterScreen from './public/RegisterScreen';
 import PartnerProfile from './public/PartnerProfile';
 import ServiceStatusScreen from './public/ServicesStatus';
 import { NavigationParams } from './types';
-import Loading from './public/Loading';
+import Home from './public/Home';
 
 const RootStack = createNativeStackNavigator<NavigationParams>({
   screens: {
@@ -22,10 +22,10 @@ const RootStack = createNativeStackNavigator<NavigationParams>({
         headerShown: false,
       },
     },
-    Loading: {
-      screen: Loading,
+    Feed: {
+      screen: Feed,
       linking: {
-        path: 'loading',
+        path: 'feed',
       },
       options: {
         headerShown: false,
