+++ src/screens/public/Home/styles.ts
@@ -3,70 +3,34 @@ import { StyleSheet } from 'react-native';
 export const styles = StyleSheet.create({
   container: {
     flex: 1,
-    backgroundColor: '#e6f0fa',
-  },
-  header: {
-    backgroundColor: '#003366',
-    padding: 15,
-    alignItems: 'center',
-  },
-  headerText: {
-    fontSize: 16,
-    fontWeight: 'bold',
-    color: '#ffffff',
+    backgroundColor: '#ff7f00',
   },
   contentContainer: {
     alignItems: 'center',
+    justifyContent: 'center',
     padding: 20,
+  },
+  logo: {
+    width: 100,
+    height: 100,
+    resizeMode: 'contain',
+    marginBottom: 10,
   },
   title: {
     fontSize: 28,
     fontWeight: 'bold',
-    color: '#003366',
-    marginBottom: 20,
-  },
-  card: {
-    backgroundColor: '#ffffff',
-    borderRadius: 8,
-    padding: 15,
-    width: '100%',
-    marginBottom: 15,
-    shadowColor: '#000',
-    shadowOffset: { width: 0, height: 2 },
-    shadowOpacity: 0.1,
-    shadowRadius: 4,
-    elevation: 2,
-  },
-  cardTitle: {
-    fontSize: 18,
-    fontWeight: 'bold',
-    color: '#003366',
-    marginBottom: 5,
-  },
-  cardDescription: {
-    fontSize: 14,
-    color: '#003366',
-    marginBottom: 10,
-  },
-  cardLocation: {
-    fontSize: 12,
-    color: '#003366',
-    marginBottom: 10,
-  },
-  button: {
-    backgroundColor: '#ff7f00',
-    padding: 10,
-    borderRadius: 8,
-    alignItems: 'center',
-  },
-  buttonText: {
     color: '#ffffff',
-    fontWeight: 'bold',
+  },
+  subtitle: {
+    fontSize: 16,
+    color: '#ffffff',
+    marginBottom: 20,
   },
   footer: {
     position: 'absolute',
     bottom: 10,
     fontSize: 12,
-    color: '#003366',
+    alignSelf: 'center',
+    color: '#ffffff',
   },
 });
