--- src/screens/public/Loading/index.ts
@@ -1,1 +1,0 @@
-export { default } from './Loading';
