--- src/screens/public/Loading/styles.ts
@@ -1,36 +1,0 @@
-import { StyleSheet } from 'react-native';
-
-export const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    backgroundColor: '#ff7f00',
-  },
-  contentContainer: {
-    alignItems: 'center',
-    justifyContent: 'center',
-    padding: 20,
-  },
-  logo: {
-    width: 100,
-    height: 100,
-    resizeMode: 'contain',
-    marginBottom: 10,
-  },
-  title: {
-    fontSize: 28,
-    fontWeight: 'bold',
-    color: '#ffffff',
-  },
-  subtitle: {
-    fontSize: 16,
-    color: '#ffffff',
-    marginBottom: 20,
-  },
-  footer: {
-    position: 'absolute',
-    bottom: 10,
-    fontSize: 12,
-    alignSelf: 'center',
-    color: '#ffffff',
-  },
-});
